Actually fix 8-byte alignment

+++ source/bignum/Num.h
@@ -85,7 +85,7 @@ public:
             uint32_t* digits; // want different name than buf to catch bugs
         } big;
 
-        long long force_8byte;
+        double force_8byte;
     };
 
     #pragma pack(pop)
